requirements.txt
requests
pytest
structlog
curlify
pydantic
pyhamcrest
psycopg2
records
sqlalchemy==1.4.46
vyper-config
allure-pytest

